+++ src/main/java/com/newsstand/service/magazine/MagazineService.java
@@ -29,7 +29,7 @@ public interface MagazineService {
      * @param categoryId Id of the category to filter magazines.
      * @return           A list of magazines.
      */
-    List<Magazine> getPageByCategoryId(Long page, Long size, Long categoryId);
+    List<Magazine> getPageByCategoryId(Integer page, Integer size, Long categoryId);
 
     /**
      * This method returns a page of magazines with publisher id.
@@ -39,7 +39,7 @@ public interface MagazineService {
      * @param publisherId Id of the publisher to filter magazines.
      * @return            A list of magazines.
      */
-    List<Magazine> getPageByPublisherId(Long page, Long size, Long publisherId);
+    List<Magazine> getPageByPublisherId(Integer page, Integer size, Long publisherId);
 
     /**
      * This method returns a page of all magazines.
@@ -48,7 +48,7 @@ public interface MagazineService {
      * @param size Size of the page.
      * @return     A list of magazines.
      */
-    List<Magazine> getPage(Long page, Long size);
+    List<Magazine> getPage(Integer page, Integer size);
 
     /**
      * This method creates new magazine.
